+++ src/main/kotlin/io/github/jamalam360/database/entity/ServerConfig.kt
@@ -28,7 +28,8 @@ data class ServerModerationConfig(
     var enabled: Boolean,
     var moderatorRole: Long,
     var mutedRole: Long,
-    var threadAutoJoinRoles: MutableList<Long>
+    var threadAutoJoinRoles: MutableList<Long>,
+    var autoSaveThreads: Boolean
 )
 
 data class ServerTagsConfig(